+++ gladius/daisyui.py
@@ -45,7 +45,7 @@ class Page(Component):
                     <link href="https://cdn.jsdelivr.net/npm/daisyui@3.7.7/dist/full.css" rel="stylesheet" type="text/css" />
 
                     <!-- tailwind -->
-                    <script src="https://cdn.tailwindcss.com"></script>
+                    <script src="https://cdn.tailwindcss.com?plugins=forms,typography,aspect-ratio,line-clamp"></script>
 
                     <!-- htmx -->
                     <script src="https://unpkg.com/htmx.org@1.9.5"></script>