+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## v0.1.4
 
-...
+Added:
+    - HTML5: H1, H2, H3, H4, P, BlockQuote, Figure, FigCaption, Strong, Em, Code
+    - HTML5: Pre, Ol, Ul, Li, Table, THead, Tr, Th, Td, Img, Video, Source, Hr
+
+Changed:
+    - Tailwind Plugins: forms, typography, aspect-ratio, line-clamp
 
 ## v0.1.3